+++ VersionHistory.md
@@ -6,11 +6,13 @@ Describe changes here when they're committed to the `master` branch. Move them t
 
 Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in accordance with [Semantic Versioning](https://semver.org/).
 
+## Released
+
+### 3.0.0
+
 * [major] Drop support for `packagediff` for now. (Too buggy.)
 * [patch] `package` target depends on `clean` for simpler package detection.
 * [major] Support SourceLink authentication, moving settings into `SourceLinkSettings`. Now disabled by default.
-
-## Released
 
 ### 2.2.0